test: move controller addresses tests in lsp3-universal-profile.spec.ts
